% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlm.R
\name{print.mvlm}
\alias{print.mvlm}
\title{Print mvlm Object}
\usage{
\method{print}{mvlm}(x, ...)
}
\arguments{
\item{x}{Output from \code{mvlm}}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\item{p-value}{Analytical p-values for the omnibus test and each predictor}
}
\description{
\code{print} method for class \code{mvlm}
}
\author{
Daniel B. McArtor (dmcartor@nd.edu) [aut, cre]
}

