% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlm.R
\name{residuals.mvlm}
\alias{residuals.mvlm}
\title{Extract mvlm Residuals}
\usage{
\method{residuals}{mvlm}(object, ...)
}
\arguments{
\item{object}{Output from \code{mvlm}}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A data frame of residuals with the same dimension as the outcome
data passed to \code{mvlm}
}
\description{
\code{residuals} method for class \code{mvlm}.
}
\examples{
data(mvlmdata)
Y <- as.matrix(Y.mvlm)
mvlm.res <- mvlm(Y ~ Cont + Cat + Ord, data = X.mvlm)
Y.resid <- resid(mvlm.res)

}
\author{
Daniel B. McArtor (dmcartor@nd.edu) [aut, cre]
}

