% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlm.R
\docType{data}
\name{X.mvlm}
\alias{X.mvlm}
\title{Simulated predictor data to illustrate the mvlm package.}
\format{An object of class \code{data.frame} with 200 rows and 3 columns.}
\usage{
X.mvlm
}
\description{
See package vignette by calling \code{vignette('mvlm-vignette')}.
}
\keyword{datasets}

