% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlm.R
\docType{data}
\name{Y.mvlm}
\alias{Y.mvlm}
\title{Simulated outcome data to illustrate the mvlm package.}
\format{An object of class \code{matrix} with 200 rows and 5 columns.}
\usage{
Y.mvlm
}
\description{
See package vignette by calling \code{vignette('mvlm-vignette')}.
}
\keyword{datasets}

